validando fluxo de resposta para caso nao encontre o arquivo no banco de dados, nos metodos, store, update e destroy.

+++ app/Http/Controllers/MarcaController.php
@@ -4,9 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Models\Marca;
 use Illuminate\Http\Request;
+use Ramsey\Uuid\Type\Integer;
 
 class MarcaController extends Controller
 {
+    public function __construct(Marca $marca){
+        $this->marca = $marca;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +19,8 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        $marcas = Marca::all();
+        //$marcas = Marca::all();
+        $marcas = $this->marca->all();
         return $marcas;
     }
 
@@ -28,18 +34,23 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca = Marca::create($request->all());
+        //$marca = Marca::create($request->all());
+        $marca = $this->marca->create($request->all());
         return $marca;
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
-     * @return \Illuminate\Http\Response
+     * @param  Integer
+     * @return string[]
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -48,11 +59,16 @@ class MarcaController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
-     * @return \Illuminate\Http\Response
+     * @param  Integer
+     * @return string[]
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+        //$marca->update($request->all());
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossivel realiza a atualizacao. O recurso solicitado não existe'];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -60,11 +76,16 @@ class MarcaController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
-     * @return \Illuminate\Http\Response
+     * @param  Integer
+     * @return string[]
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        //$marca->delete();
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossivel realiza a exclusao. O recurso solicitado não existe'];
+        }
         $marca->delete();
         return ['msg' => "A marca foi removida com sucesso"];
     }
